requirements.txt
langchain>=0.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
tavily-python==0.5.4
rich>=13.0.0
flask>=2.0.0 